Extracting runtime file.
