+++ zod-plugin/index.ts
@@ -1,2 +1,3 @@
-import "./augmentation"; // side effects here
+import "./augmentation";
+import "./runtime"; // side effects here
 export { getBrand } from "./helpers";
