+++ zod-plugin/vitest.setup.ts
@@ -1,1 +1,1 @@
-import "./augmentation";
+import "./runtime";
